added auth middle ware and schema methods for hashing and generating token

+++ server/src/routers/users.js
@@ -1,6 +1,5 @@
 import express from "express";
 import _ from "lodash";
-import jwt from "jsonwebtoken";
 import bcrypt from "bcrypt";
 import {
   userLoginSchema,
@@ -10,8 +9,12 @@ import {
 import validate from "../middleware/joiValidation.js";
 import usersModel from "../models/usersModel.js";
 import UserModel from "../models/usersModel.js";
+import dotenv from "dotenv";
+import auth from "../middleware/auth.js";
 
 const router = express.Router();
+
+dotenv.config();
 
 // get all users
 
@@ -37,13 +40,14 @@ router.post("/register", validate(userRegistrationSchema), async (req, res) => {
       _.pick(req.body, ["name", "password", "email", "phone", "gender"])
     );
 
-    const salt = await bcrypt.genSalt(10);
-    const hashPassword = await bcrypt.hash(req.body.password, salt);
-    newUser.password = hashPassword;
+    newUser.password = await newUser.hashPassword(req.body.password)
 
     await newUser.save();
 
+    const token = newUser.generateAuthToken()
+
     return res
+      .header("x-auth-token", token)
       .status(200)
       .send(_.pick(newUser, ["name", "email", "phone", "gender"]));
   } catch (err) {
@@ -66,19 +70,9 @@ router.post("/login", validate(userLoginSchema), async (req, res) => {
     if (!validPassword)
       return res.status(400).send("Invalid email or password");
 
-    const token = jwt.sign(
-      {
-        name: user.name,
-        email: user.email,
-        phone: user.phone,
-        gender: user.gender,
-        _id: user._id,
-        role: user.role,
-      },
-      "e3d5f00120453f0cef0946af3eed5c7f421da013849fbba83107a4a91bde1ec8244ea00f4354cc69cacb490e16e700912305d018c0172734d7708f58261799c0"
-    );
+    const token = user.generateAuthToken()
 
-    return res.status(200).send(token);
+    return res.status(200).header("x-auth-token", token).send(_.pick(user, "_id"));
     // return res.status(200).send(_.pick(user, ['name', 'email', 'phone', 'gender', 'image', '_id', 'role']))
   } catch (err) {
     return res.status(500).send(err);
@@ -87,7 +81,7 @@ router.post("/login", validate(userLoginSchema), async (req, res) => {
 
 // edit profile
 
-router.put("/edit/:id", validate(userUpdateSchema), async (req, res) => {
+router.put("/edit/:id", auth, validate(userUpdateSchema), async (req, res) => {
   try {
     const id = req.params.id;
     const data = _.pick(req.body, [
@@ -97,15 +91,21 @@ router.put("/edit/:id", validate(userUpdateSchema), async (req, res) => {
       "phone",
       "image",
     ]);
-    if (data.password) {
-      const salt = await bcrypt.genSalt(10);
-      data.password = await bcrypt.hash(data.password, salt);
-    }
+
+    if(req.user._id !== id) return res.status(401).send("unauthorized access")
+
     const user = await usersModel.findByIdAndUpdate(id, data, { new: true });
+
+    if (data.password) {
+      data.password = await user.hashPassword(data.password)
+    }
+
     if (!user) return res.status(400).send("not found");
+
     return res
       .status(200)
       .send(_.pick(user, ["name", "email", "phone", "image", "_id"]));
+
   } catch (err) {
     return res.status(500).send(err);
   }
@@ -113,9 +113,11 @@ router.put("/edit/:id", validate(userUpdateSchema), async (req, res) => {
 
 // delete user
 
-router.delete("/delete/:id", async (req, res) => {
+router.delete("/delete/:id", auth, async (req, res) => {
   try {
     const id = req.params.id;
+    if(req.user._id !== id) return res.status(401).send("unauthorized access")
+      
     const user = await UserModel.findByIdAndDelete(id);
     if (!user) return res.status(400).send("not found");
     return res.status(200).send(_.pick(user, ["_id", "name", "email"]));
